Separate socket from channels
